requirements.txt
# A list of dependencies.
Django>=4.0.5,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<3
drf-spectacular>=0.22.1,<0.23
django-phonenumber-field[phonenumbers]

